Emit all "asymptotically equivalent" plans
This commit also has loads of sweet sweet optimizations to make the
synthesis much faster. It's slower than it once was, but not by much.
This is good, considering we're doing a lot more work now!

Also cleaned up stupid one-element lists (workaround for Python 2.x not
having the "nonlocal" keyword).