Added --cpp-abstract-record option
This can make the refactoring job a lot easier when introducing a
synthesized data structure. This way, the client is responsible for
telling the synthesizer how to access the fields on the record type,
and can store them in whatever way is most convenient.

In particular, this will make the Bullet refactoring much, much easier.