Simplified acceleration rules
Since I introduced bag subtraction and implemented incrementalization in terms
of that, the acceleration rule for filter-out-singleton is no longer useful.
Instead, a simpler rule that breaks functions of bag additions and subtractions
suffices.

This rule looks at any expression containing a bag addition or subtraction, and
turns it into that same expression over each side. This allows complex things
like sums-of-maps-of-filters-of-bag-subtractions to be split into their
constituent parts.