Disable disjunctive filter breaking by default
Previously, Cozy had an always-on desugaring rule that says

    Filter (P or Q) xs ---> (Filter P xs) + (Filter (not P and Q) xs).

This improves performance in some cases, but harms it in others. Furthermore it
complicates the code and would need to be explained in any potential paper, so
I am disabling it. It is not a clear win. Depending on how the other case
studies turn out, I may add this back as an acceleration rule (possibly with
some other intelligent logic).