Improvements to fixup_binders
Two big things:
  1. Top-down rewriting
  2. Dynamically grow the pool of available binders

The first is necessary because subst() may introduce additional binders to do
capture-avoiding substitution.

The second is necessary because incrementalization may produce queries with
extra binders, and we need to handle it properly. This may be able to go away
once the code in high_level_interface.py selects the appropriate number of
binders out-of-the-box.