Instantiate binders based on evaluation of target expression
This is a very important change. Previously, Cozy would try to guess at what
values could possibly flow into any given binder, and instantiate examples that
way. However, that heuristic was woefully incomplete. Consider:

    Sum (Filter {\b -> b > 0} (Map {\_ -> 1} xs)).

In that expression, the binder "b" will always be bound to the value "1",
regardless of the contents of xs. Thus, unless xs contains the value "1", the
old instantiation heuristic would give a lot of incorrect values for "b".

The new heuristic instantiates based on all values that *actually* flow into
"b" based on the target expression's behavior on the existing examples. This is
much more likely to be correct, and I suspect it keeps the number of examples
smaller as well in many cases.