EStateVar for explicit boundaries between runtime & state expressions
This is part one of my two-step plan to permanently eliminate monotonicity
violations from the cost model. The issue is that expressions may have one cost
when discovered and inserted into the cache, but then take on another once they
are "in-context" as part of another expression. To fix this, we'll need the
synthesizer to reason explicitly about this boundary.