Fixed a bug where unused query arguments could cause errors
The problem is that the Builder assembled in high_level_synthesis.py
might use those irrelevant argument variables in built expressions, but
the algorithm in core.py does not produce counterexamples involving
those variables. This mismatch could lead to core.py trying to
fingerprint expressions with free variables not in any example, which
in turn leads to problems.

The fix is to simply prevent the Builder from constructing such
expressions. In rare cases this may improve synthesis speeds as well,
since there are fewer variables to build expressions out of.