Update and improve the readme
In addition to adding more useful text to the readme, this commit also
updates the recommended Z3 version to 4.5, the latest stable release.
Previously the readme suggested a weird formerly-bleeding-edge version
of Z3 that was supported almost nowhere (in order to avoid crashes in
the Python 3 bindings). The 4.5 version is newer, just as stable, and
actually available in some package managers.