Fixed a set of overly-aggressive assumptions
In the expression

    Filter {\x -> P} Q

where x is free in Q, you cannot assume that x is in Q while evaluating
P. The name clash will mess you up!

We can make a safe assumption by alpha-renaming x to something else in
the filter predicate, but this would introduce a new name and would
wreak havoc on the synthesis algorithm (since it assumes that it knows
all names in advance).

The safest option is to just avoid generating this assumption when it
may be invalid. This could degrade the quality of the synthesis output,
but at least it is correct!