Optimized _possible_replacements in core.py
This has been a long time coming. Previously, _possible_replacements
ran in O(E*W*V) time, where E is the number of instantiated examples
(often 500+), W is the number of watched expressions (i.e. the number
of nodes in target expression, often 20-30), and V is the time required
to evaluate the new candidate on an example PLUS the time required
to evaluate each watched expression on an example.

What a mess!

The new approach indexes watched expressions in a structure along with
their assumptions. It allows us to look up candidate watched expressions
based only on the fingerprint of the target without evaluating anything,
and has runtime O(E) where E is the number of examples.