Symbolic cost model
This commit makes the cost model symbolic in terms of set cardinalities.
The mechanism is quite simple:

 - an expression's cost is an integer-valued expression that may involve
   unknown variables representing the cardinalities of various sets
 - cardinalities are ordered using containment:
     if forall x, x in A --> x in B,
     then |A| >= |B|.
 - costs are partially ordered: a symbolic cost might be always better,
   always at least as good, equivalent, or sometimes better sometimes
   worse relative to another cost.

This is a large commit, and one that I have been on the fence about for
some time. However, it allows us to remove a great deal of hackery in
the cost model. Furthermore, it does not substantially impact
performance, since the previous cost model also made solver calls.