Introduced "deep equality"
I hate the idea of having multiple definitions of equality, but this is
really necessary since regular double-equals "equality" does not mean
"substitutability" in Cozy. By that I mean:

    (x == y)  =/=>  (f(x) == f(y))

Example:

    EWithAlteredValue(a, 0) == EWithAlteredValue(a, 1)
    but
    EWithAlteredValue(a, 0).val != EWithAlteredValue(a, 1).val

This creates serious problems for synthesis correctness, which
previously used double-equals equality. This could lead to situations
where the "value" part of EWithAlteredValue expressions got altered,
leading to incorrect results.

I have fixed the problem by adding a second, internal-only notion of
equality (triple-equals, "===") defined as

    forall x y f,
    (x === y)  ==>  (f(x) == f(y))

That is, all observable properties of x and y are identical when they
are deep-equal.