Implemented checking for well-formed "the" expressions
The "the" expression extracts just the single element in a singleton
collection or nothing if the collection is empty.

It is illegal to use "the" on a collection that may have more than one
element, since its behavior is undefined in that case.

The synthesis backend already avoids generating "the" expressions on
non-singleton lists.