Massive improvement to example instantiation
It turns out that we do not need any fishy example-duplication logic.
Instead, we can just ask the solver to produce a counterexample that
differentiates BOTH the top-level expressions AND the subexpressions.

Besides fixing some "duplicate counterexample" errors I've been seeing,
this also improves performance by A LOT, since the number of
instantiated examples is orders-of-magnitude smaller.