Moved handle address-value relationship handling to preprocessing step
Due to the existence of EWithAlteredValue, it no longer makes sense for
the solver to assume that two handles with the same address necessarily
have the same value. Instead, that is a sort of "implicit assumption"
at the top of every method on the data structure. We know that
assumption is always true since that's how heap-allocated objects work:
if they have the same address, then they have the same value.