Individual per-expression costs in synthesis cache
This change
  1. introduces a dedicated SeenSet type for managing seen expressions
  2. uses OrderedDict to reduce nondeterminacy in the seen set
  3. tracks cost of each expression individually

The core component of this commit---item 3---is a good idea because our
symbolic cost model is only partially-ordered.  Consider this scenario:

  known expressions:  A   B   C
                       \ /
  new expression:       X

  such that
    X is better than A and better than B but
    X is unordered with respect to C

The old code would have evicted all known expressions in favor of X,
thus dropping the potentially-useful C.  The new code evicts only A and
B, keeping X and C.  The invariant is that all expressions cached for
a given fingerprint are unordered w.r.t. each other.

This change makes performance slightly worse, but the hope is that
synthesis now catches some good expressions it might have otherwise
missed.