Counterexample-guided cost model
This commit greatly improves the performance of the cost model by using
examples to perform cost comparisons.  As an added benefit, this model
is rich enough to capture the fact that conditionals only execute one
branch and that boolean connectives short-circuit.

This commit also adds a verbose mode, enabled with the --verbose flag.
A few core parts of the code use the new logging infrastructure so you
can see where Cozy is spending its time.