Greatly improved encoding times
A major bottleneck in Cozy is how long it takes to convert from Cozy
expressions to Z3 formulas.  The problem is caused by needing to
re-traverse the body of every lambda for every bag element it could be
applied to.  When bags are expanded out to size 4 (the current default)
this results in many, many unnecessary calls, especially if lambdas
are deeply nested.

This commit addresses the problem by using Z3's builtin substitution
function to do lambda substitution AFTER encoding.  Therefore we can
traverse the entire AST in a single pass.  Doing the substitution still
incurs some penalty, but it seems to be much less severe in practice.