Moved a lot of heuristics to a better place
The code in wf.py was always intended to be for enforing well-formedness of
EStateVar usage.  It happened to be a convenient place to insert hard filters
to ignore classes of expressions, and so it morphed into a dual-purpose
function that enforced well-formedness and miscellaneous heuristics.

This commit moves the heuristics to core.py, where they are actually used, and
reduces wf.py to only checking well-formedness.  It also introduces a few
internal options to control some of the heuristics.