Fixed a subtle solver encoding bug surrounding ESorted
This commit fixes a bug where expressions containing ESorted inside an
ELambda expression would be solved incorrectly.  Since the encoding for
ESorted introduces a fresh variable, that variable needs to be
introduced every time the function would be called.  However, there was
previously an optimization in the encoding for ELambda expressions to
deliberately AVOID re-encoding the body of an ELambda every time it
would be called.

The fix here sacrifices performance for correctness, replacing the
encoding for ELambda with the naive version that works as expected.  A
faster implementation would be possible if the encoding for ELambda
captured all variables introduced while encoding the body and
re-instantiated them.  However, in this case I do not feel the added
complexity is worth the minor performance gain.

This commit also introduces two unit tests for the solver that include
ESorted.  The second of these exposes the bug fixed in this commit.