Added counter for number of improvements done
This counter is added to aid with evaluating the rate at which improvements are
being made to implementations. This might be useful for determining speed-up improvements for Cozy. 

Right now it is incremented every time [step #1](https://github.com/CozySynthesizer/cozy/blob/master/cozy/synthesis/core.py#L208) of the
`improve` loop finishes.