Close a possible race condition
There is a gap between when a job starts and when it installs the
signal handler for SIGINT.  It is actually quite common for Cozy to
stop a job during the gap; the it happens while running the unit tests,
for instance.

The gap is an unfortunate necessity; Python does not offer a way to
atomically start a new process with a given signal handler
pre-installed without using the now-discouraged "fork" multiprocessing
method.

The gap is closed awkwardly: when a job is requested to stop, it only
sends SIGINT if it has evidence that the child has already set itself
up for receiving that signal.  In the interim period it will only set
the stop flag.  The only problem with this solution is that any Z3
calls that happen before setting up the signal handler will not be
interrupted---but fortunately, today, there aren't any.