Forward compatibility: use "spawn" method instead of "fork"
The "spawn" method is already the default on Windows and will be the
default on MacOS starting with Python 3.8.  Switching to this method
required a few tweaks since SafeQueue objects can't be pickled.

There is still a lingering problem: since signal handlers are not
inherited by the child, there is a gap between when the process starts
and when the signal handler is installed where a SIGINT can cause an
exception instead of a graceful stop.