Restore sys.stdout even when an exception is thrown
This resolves the annoyting "ValueError: I/O operation on closed file"
errors appearing in Travis.  The errors are harmless and specific to
Python 3.5, but they look concerning and they muddy Cozy's output.

Aside: these errors are not fixed in Python 3.6+, but they are hidden
because there are other common situations where they happen.  See:
 - https://bugs.python.org/issue28326
 - https://bugs.python.org/issue31804
 - https://github.com/python/cpython/commit/e756f66c83786ee82f5f7d45931ae50a6931dd7f